+++ frontend/src/pages/callPage.tsx
@@ -1,14 +1,14 @@
 import { FC } from "react";
 import styled from "styled-components";
-import Circle from "../components/layout/Cricle";
+import { CallCircle } from "../utils/CallCricle";
 
 const CallPageContainer = styled.div`
-  text-align: center;
-  overflow: hidden;
+    text-align: center;
+    overflow: hidden;
 `;
 
 const PassButton = styled.div`
-  margin-top: 2rem;
+    margin-top: 2rem;
 `;
 
 const EndButton = styled.div``;
@@ -18,7 +18,7 @@ const CallPage: FC = () => {
   return (
     <CallPageContainer>
       <PassButton>パス</PassButton>
-      <Circle message={"Call"} onClick={onCalling} />
+      <CallCircle onClick={onCalling}/>
       <EndButton>完了</EndButton>
     </CallPageContainer>
   );
