+++ frontend/src/pages/startPage.tsx
@@ -1,23 +1,23 @@
 import { FC } from "react";
 import styled from "styled-components";
-import Button from "../components/layout/Button";
-import Cricle from "../components/layout/Cricle";
+import { Button } from "../utils/Button";
+import { MessageCricle } from "../utils/MessageCricle";
 
 const StartPageContainer = styled.div`
-  text-align: center;
-  overflow: hidden;
+    text-align: center;
+    overflow: hidden;
 `;
 
 const StartPage: FC = () => {
-  const onStart = () => {};
-  return (
-    <StartPageContainer>
-      <div>start page</div>
-      <Cricle message={"appName"} onClick={() => {}} />
-      <div>日付</div>
-      <Button message={"営業開始"} onClick={onStart} />
-    </StartPageContainer>
-  );
+    const onStart = () => {};
+    return (
+        <StartPageContainer>
+            <div>start page</div>
+            <MessageCricle message={"appName"} />
+            <div>日付</div>
+            <Button message={"営業開始"} onClick={onStart} />
+        </StartPageContainer>
+    );
 };
 
 export default StartPage;
